Fixed Credential Issue
Fixed an issue where AD wasn't accepting the user password because it wasn't a secure string. Also fixed the issue where I took admin credential input but then didn't pass it to any of the script that was interacting with AD so it was just trying and failing with the account who ran the script. TODO: Test all options of this script and then I believe it's ready for production.